# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram Bingo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /users/{id}/balance:
    patch:
      operationId: adjustBalance
      tags: [users]
      summary: Adjust user balance (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BalanceAdjustInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}/ban:
    patch:
      operationId: banUser
      tags: [users]
      summary: Ban or unban user (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanUserInput"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── GAMES ──────────────────────────────────────────────────────────────────
  /games:
    get:
      operationId: listGames
      tags: [games]
      summary: List games
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 10 }
        - name: status
          in: query
          schema:
            type: string
            enum: [waiting, active, finished]
      responses:
        "200":
          description: Paginated games
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GamePage"
  /games/current:
    get:
      operationId: getCurrentGame
      tags: [games]
      summary: Get current active or waiting game
      responses:
        "200":
          description: Current game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: No active game
  /games/{id}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get game by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Game
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Game"
        "404":
          description: Not found
  /games/{id}/tickets:
    post:
      operationId: buyTicket
      tags: [tickets]
      summary: Buy a ticket for a game
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Ticket purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ticket"
        "400":
          description: Game not accepting tickets or insufficient balance
    get:
      operationId: getMyGameTickets
      tags: [tickets]
      summary: Get current user tickets for a game
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ticket"
  /games/{id}/claim:
    post:
      operationId: claimBingo
      tags: [games]
      summary: Claim bingo for a game
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BingoClaimInput"
      responses:
        "200":
          description: Claim result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BingoClaimResult"
        "400":
          description: Invalid claim

  # ─── TICKETS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, telegramId, username, firstName, balance, role, isBanned, createdAt]
      properties:
        id: { type: integer }
        telegramId: { type: string }
        username: { type: string }
        firstName: { type: string }
        lastName: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        balance: { type: number }
        totalWinnings: { type: number }
        totalGames: { type: integer }
        totalWins: { type: integer }
        role: { type: string, enum: [player, admin] }
        isBanned: { type: boolean }
        referralCode: { type: string }
        referredBy: { type: ["integer", "null"] }
        createdAt: { type: string, format: date-time }
    UserPage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    BalanceAdjustInput:
      type: object
      required: [amount, reason]
      properties:
        amount: { type: number }
        reason: { type: string }
    BanUserInput:
      type: object
      required: [banned]
      properties:
        banned: { type: boolean }
...
    Game:
      type: object
      required: [id, status, ticketPrice, totalPot, houseFee, prizePool, houseEarnings, drawnNumbers, playerCount, createdAt]
      properties:
        id: { type: integer }
        status: { type: string, enum: [waiting, active, finished] }
        ticketPrice: { type: number }
        totalPot: { type: number }
        houseFee: { type: number }
        prizePool: { type: number }
        houseEarnings: { type: number }
        drawnNumbers: { type: array, items: { type: integer } }
        playerCount: { type: integer }
        winnerId: { type: ["integer", "null"] }
        winnerUsername: { type: ["string", "null"] }
        winPattern: { type: ["string", "null"] }
        startedAt: { type: ["string", "null"], format: date-time }
        endedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    GamePage:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Game"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }

    # ─── TICKET ───────────────────────────────────────────────────────────────
    Ticket:
      type: object
      required: [id, gameId, userId, card, markedNumbers, isWinner, purchasedAt]
      properties:
        id: { type: integer }
        gameId: { type: integer }
        userId: { type: integer }
        card:
          type: array
          items:
            type: array
            items: { type: integer }
        markedNumbers:
          type: array
          items: { type: integer }
        isWinner: { type: boolean }
        prizeAmount: { type: ["number", "null"] }
        purchasedAt: { type: string, format: date-time }
...
    BingoClaimInput:
      type: object
      required: [ticketId]
      properties:
        ticketId: { type: integer }
    BingoClaimResult:
      type: object
      required: [valid, message]
      properties:
        valid: { type: boolean }
        message: { type: string }
        prizeAmount: { type: ["number", "null"] }
        pattern: { type: ["string", "null"] }

    # ─── DEPOSIT ──────────────────────────────────────────────────────────────
...