# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Telegram Bingo API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from './gameStatus';

export interface Game {
  id: number;
  status: GameStatus;
  ticketPrice: number;
  totalPot: number;
  houseFee: number;
  prizePool: number;
  houseEarnings: number;
  drawnNumbers: number[];
  playerCount: number;
  /** @nullable */
  winnerId?: number | null;
  /** @nullable */
  winnerUsername?: string | null;
  /** @nullable */
  winPattern?: string | null;
  /** @nullable */
  startedAt?: Date | null;
  /** @nullable */
  endedAt?: Date | null;
  createdAt: Date;
}